Moved OpenclMatrix to new file
